include dynamic plugin prop values in plugin detail ajax response
